+++ src/main/java/amara/command/AddToDoCommand.java
@@ -1,9 +1,15 @@
+package amara.command;
 import java.util.ArrayList;
+
+import amara.storage.Storage;
+import amara.task.Task;
+import amara.task.ToDo;
+import amara.ui.Ui;
 
 public class AddToDoCommand extends Command {
     private ToDo todo;
 
-    AddToDoCommand(ToDo todo) {
+    public AddToDoCommand(ToDo todo) {
         this.todo = todo;
     }
