+++ src/main/java/amara/command/ByeCommand.java
@@ -1,4 +1,10 @@
+package amara.command;
 import java.util.ArrayList;
+
+import amara.exceptions.AmaraException;
+import amara.storage.Storage;
+import amara.task.Task;
+import amara.ui.Ui;
 
 public class ByeCommand extends Command {
     @Override
