+++ src/main/java/amara/command/Command.java
@@ -1,4 +1,10 @@
+package amara.command;
 import java.util.ArrayList;
+
+import amara.exceptions.AmaraException;
+import amara.storage.Storage;
+import amara.task.Task;
+import amara.ui.Ui;
 
 public abstract class Command {
     public boolean isBye() {
